install-docker support install from an rpm

+++ scripts/config-yum.sh
@@ -1,8 +1,9 @@
 echo "add yum repo"
 sed -i 's/enabled=1/enabled=0/g' /etc/yum/pluginconf.d/fastestmirror.conf
-rm -f /etc/yum.repos.d/CentOS-Base.repo
-curl ${YUM_BASE} -o /etc/yum.repos.d/CentOS-Base.repo
-
+if [[ ! -z "${YUM_YUM_BASE}" ]]; then
+  rm -f /etc/yum.repos.d/CentOS-Base.repo
+  curl ${YUM_BASE} -o /etc/yum.repos.d/CentOS-Base.repo
+fi
 if [[ ! -z "${YUM_DOCKER}" ]]; then
   yum-config-manager --add-repo ${YUM_DOCKER}
 fi
